feat(socket): Add `socket` example
This example just showcases how `socket` is used in the most basic sense.

It is not a simple app example like `showip`, that is why the received socket file descriptor is just printed to the output.

Along with this change, the full title of `getaddrinfo` section is added to the project.

Signed-off-by: acikgozb <acikgozb@proton.me>