feat(recv): Add `recv` example
Since there were no examples about `recv()`, I added a basic example to
showcase the usage.

This example showcases a `recv()` call where it tries to read `len`
bytes from an incoming message, which is specified as 30.

Contrary from other examples, the message received from an incoming
connection is not written with `println!`. It is written with
`io::stdout()` to not deal with any encoding at all.

Signed-off-by: acikgozb <acikgozb@proton.me>