fix(accept): Fix SAFETY comment in `accept` example
During the `sendto` example, I've noticed that when we use our local host as a `node` for `getaddrinfo()`, the resulted IP address is actually the loopback IP address based on the address family provided in `hints` (IP or IPv6).

It is NOT our local host's public IP address. Therefore, the safety comment in `accept()` is fixed.

Signed-off-by: acikgozb <acikgozb@proton.me>